requirements.txt
absl-py==1.3.0
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
blinker==1.6.2
Bottleneck==1.3.5
cachetools==5.2.0
category-encoders==1.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cornac==1.14.2
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-login-required-middleware==0.9.0
django-parler==2.3
django-phrase==1.1.0
django-redis==5.2.0
django-rosetta==0.9.9
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.0.4
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.11.1
hypothesis==6.61.0
idna==3.4
importlib-metadata==5.2.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
kombu==5.2.4
libclang==14.0.6
lightfm==1.16
lightgbm==3.3.3
llvmlite==0.39.1
mando==0.6.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
memory-profiler==0.61.0
mpmath==1.2.1
mypy-extensions==0.4.3
nltk==3.8
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandera==0.13.4
patsy==0.5.3
Pillow==9.3.0
polib==1.2.0
powerlaw==1.5
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.7
PyYAML==5.4.1
radon==5.1.0
recommenders==1.1.1
redis==4.5.5
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rosetta==0.3
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scikit-surprise==1.1.3
scipy==1.9.3
seaborn==0.12.1
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
statsmodels==0.13.5
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
tmdbsimple==2.9.1
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.3.2
wrapt==1.14.1
zipp==3.11.0
